Capstone Project/notebook-lm/requirements.txt
streamlit
langchain
langchain-ollama
langchain-community
langchain-chroma
langchain-text-splitters
langgraph
chromadb
pypdf
tavily-python
python-dotenv
requests
